pyproject.toml
[project]
name = "carteles"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.9",
    "pydantic>=2.6",
    "Pillow>=10.2",
    "openai>=1.40",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "numpy>=1.24",
    "pyjwt>=2.8",
]

[tool.uv]
dev-dependencies = []
